#120: closeKeyboard() replaced with closeSoftKeyboard(). Matcher class extracted. Sleep removed, will see what Codacy will tell about that.
